+++ submission/submission-core/src/main/kotlin/ac/uk/ebi/biostd/submission/config/JmsPublishingConfig.kt
@@ -1,4 +1,4 @@
-package ac.uk.ebi.biostd.common.config
+package ac.uk.ebi.biostd.submission.config
 
 import ac.uk.ebi.biostd.common.properties.ApplicationProperties
 import org.springframework.amqp.rabbit.connection.ConnectionFactory
