+++ submission/submission-core/src/main/kotlin/ac/uk/ebi/biostd/submission/config/SecurityConfig.kt
@@ -3,17 +3,10 @@ package ac.uk.ebi.biostd.submission.config
 import ac.uk.ebi.biostd.common.properties.ApplicationProperties
 import ac.uk.ebi.biostd.persistence.common.service.SubmissionMetaQueryService
 import ac.uk.ebi.biostd.persistence.common.service.UserPermissionsService
-import ac.uk.ebi.biostd.persistence.repositories.AccessPermissionRepository
 import ac.uk.ebi.biostd.persistence.repositories.AccessTagDataRepo
 import ac.uk.ebi.biostd.persistence.repositories.TokenDataRepository
 import ac.uk.ebi.biostd.persistence.repositories.UserDataRepository
 import ac.uk.ebi.biostd.persistence.repositories.UserGroupDataRepository
-import ac.uk.ebi.biostd.security.domain.service.ExtUserService
-import ac.uk.ebi.biostd.security.domain.service.PermissionService
-import ac.uk.ebi.biostd.security.web.SecurityMapper
-import ac.uk.ebi.biostd.security.web.exception.SecurityAccessDeniedHandler
-import ac.uk.ebi.biostd.security.web.exception.SecurityAuthEntryPoint
-import com.fasterxml.jackson.databind.ObjectMapper
 import ebi.ac.uk.security.integration.SecurityModuleConfig
 import ebi.ac.uk.security.integration.components.IGroupService
 import ebi.ac.uk.security.integration.components.ISecurityFilter
@@ -28,17 +21,8 @@ import uk.ac.ebi.events.service.EventsPublisherService
 @Configuration
 @Import(FilePersistenceConfig::class)
 @Suppress("TooManyFunctions")
-class SecurityConfig(private val objectMapper: ObjectMapper, properties: ApplicationProperties) {
+class SecurityConfig(properties: ApplicationProperties) {
     private val securityProps = properties.security
-
-    @Bean
-    fun securityMapper() = SecurityMapper()
-
-    @Bean
-    fun accessDeniedHandler(): SecurityAccessDeniedHandler = SecurityAccessDeniedHandler(objectMapper)
-
-    @Bean
-    fun securityAuthenticationHandler(): SecurityAuthEntryPoint = SecurityAuthEntryPoint(objectMapper)
 
     @Bean
     @SuppressWarnings("LongParameterList")
@@ -78,14 +62,4 @@ class SecurityConfig(private val objectMapper: ObjectMapper, properties: Applica
 
     @Bean
     fun securityFilter(securityConfig: SecurityModuleConfig): ISecurityFilter = securityConfig.securityFilter()
-
-    @Bean
-    fun extUserService(userDataRepository: UserDataRepository): ExtUserService = ExtUserService(userDataRepository)
-
-    @Bean
-    fun permissionService(
-        permissionRepository: AccessPermissionRepository,
-        userDataRepository: UserDataRepository,
-        accessTagDataRepository: AccessTagDataRepo,
-    ) = PermissionService(permissionRepository, userDataRepository, accessTagDataRepository)
 }
