+++ submission/submission-core/src/main/kotlin/ac/uk/ebi/biostd/submission/config/SubmissionConfig.kt
@@ -1,30 +1,18 @@
-package ac.uk.ebi.biostd.common.config
+package ac.uk.ebi.biostd.submission.config
 
-import ac.uk.ebi.biostd.files.service.FileServiceFactory
 import ac.uk.ebi.biostd.integration.SerializationService
-import ac.uk.ebi.biostd.persistence.common.service.CollectionDataService
-import ac.uk.ebi.biostd.persistence.common.service.SubmissionDraftPersistenceService
 import ac.uk.ebi.biostd.persistence.common.service.SubmissionFilesPersistenceService
-import ac.uk.ebi.biostd.persistence.common.service.SubmissionMetaQueryService
 import ac.uk.ebi.biostd.persistence.common.service.SubmissionPersistenceQueryService
 import ac.uk.ebi.biostd.persistence.common.service.SubmissionPersistenceService
 import ac.uk.ebi.biostd.persistence.filesystem.api.FileStorageService
-import ac.uk.ebi.biostd.submission.config.FilePersistenceConfig
-import ac.uk.ebi.biostd.submission.config.SecurityConfig
 import ac.uk.ebi.biostd.submission.domain.extended.ExtSubmissionQueryService
 import ac.uk.ebi.biostd.submission.domain.extended.ExtSubmissionService
 import ac.uk.ebi.biostd.submission.domain.extended.ExtSubmissionSubmitter
-import ac.uk.ebi.biostd.submission.domain.helpers.CollectionService
-import ac.uk.ebi.biostd.submission.domain.helpers.OnBehalfUtils
 import ac.uk.ebi.biostd.submission.domain.request.SubmissionStagesHandler
-import ac.uk.ebi.biostd.submission.domain.service.SubmissionDraftService
 import ac.uk.ebi.biostd.submission.domain.submission.SubmissionQueryService
 import ac.uk.ebi.biostd.submission.domain.submission.SubmissionService
 import ac.uk.ebi.biostd.submission.domain.submission.SubmissionSubmitter
-import ac.uk.ebi.biostd.submission.service.FileSourcesService
 import ac.uk.ebi.biostd.submission.stats.SubmissionStatsService
-import ac.uk.ebi.biostd.submission.web.handlers.SubmitRequestBuilder
-import ac.uk.ebi.biostd.submission.web.handlers.SubmitWebHandler
 import ebi.ac.uk.extended.mapping.to.ToFileListMapper
 import ebi.ac.uk.extended.mapping.to.ToSubmissionMapper
 import ebi.ac.uk.security.integration.components.ISecurityQueryService
@@ -100,58 +88,4 @@ class SubmissionConfig {
             securityQueryService,
             eventsPublisherService,
         )
-
-    @Bean
-    fun projectService(
-        collectionSqlDataService: CollectionDataService,
-        userPrivilegeService: IUserPrivilegesService,
-    ): CollectionService = CollectionService(collectionSqlDataService, userPrivilegeService)
-
-    @Bean
-    fun submitHandler(
-        submissionService: SubmissionService,
-        extSubmissionQueryService: ExtSubmissionQueryService,
-        fileSourcesService: FileSourcesService,
-        serializationService: SerializationService,
-        toSubmissionMapper: ToSubmissionMapper,
-        queryService: SubmissionMetaQueryService,
-        fileServiceFactory: FileServiceFactory,
-    ): SubmitWebHandler =
-        SubmitWebHandler(
-            submissionService,
-            extSubmissionQueryService,
-            fileSourcesService,
-            serializationService,
-            toSubmissionMapper,
-            queryService,
-            fileServiceFactory
-        )
-
-    @Bean
-    fun submissionDraftService(
-        submitWebHandler: SubmitWebHandler,
-        toSubmissionMapper: ToSubmissionMapper,
-        serializationService: SerializationService,
-        submitRequestBuilder: SubmitRequestBuilder,
-        userPrivilegesService: IUserPrivilegesService,
-        submissionQueryService: SubmissionPersistenceQueryService,
-        persistenceDraftService: SubmissionDraftPersistenceService,
-    ): SubmissionDraftService =
-        SubmissionDraftService(
-            submitWebHandler,
-            toSubmissionMapper,
-            serializationService,
-            submitRequestBuilder,
-            userPrivilegesService,
-            submissionQueryService,
-            persistenceDraftService,
-        )
-
-    @Bean
-    fun submitRequestBuilder(
-        onBehalfUtils: OnBehalfUtils,
-    ): SubmitRequestBuilder = SubmitRequestBuilder(onBehalfUtils)
-
-    @Bean
-    fun onBehalfUtils(securityQueryService: ISecurityQueryService): OnBehalfUtils = OnBehalfUtils(securityQueryService)
 }
