+++ submission/submission-webapp/src/main/kotlin/ac/uk/ebi/biostd/common/config/internal/SubmissionWebConfig.kt
@@ -1,16 +1,79 @@
 package ac.uk.ebi.biostd.common.config.internal
 
 import ac.uk.ebi.biostd.common.properties.ApplicationProperties
+import ac.uk.ebi.biostd.files.service.FileServiceFactory
+import ac.uk.ebi.biostd.integration.SerializationService
+import ac.uk.ebi.biostd.persistence.common.service.CollectionDataService
+import ac.uk.ebi.biostd.persistence.common.service.SubmissionDraftPersistenceService
+import ac.uk.ebi.biostd.persistence.common.service.SubmissionMetaQueryService
+import ac.uk.ebi.biostd.persistence.common.service.SubmissionPersistenceQueryService
+import ac.uk.ebi.biostd.submission.domain.extended.ExtSubmissionQueryService
+import ac.uk.ebi.biostd.submission.domain.helpers.CollectionService
+import ac.uk.ebi.biostd.submission.domain.helpers.OnBehalfUtils
+import ac.uk.ebi.biostd.submission.domain.service.SubmissionDraftService
 import ac.uk.ebi.biostd.submission.domain.submission.SubmissionQueryService
 import ac.uk.ebi.biostd.submission.domain.submission.SubmissionService
+import ac.uk.ebi.biostd.submission.service.FileSourcesService
 import ac.uk.ebi.biostd.submission.web.handlers.SubmissionsWebHandler
+import ac.uk.ebi.biostd.submission.web.handlers.SubmitRequestBuilder
+import ac.uk.ebi.biostd.submission.web.handlers.SubmitWebHandler
 import ac.uk.ebi.biostd.submission.web.resources.ext.ExtendedPageMapper
+import ebi.ac.uk.extended.mapping.to.ToSubmissionMapper
+import ebi.ac.uk.security.integration.components.ISecurityQueryService
+import ebi.ac.uk.security.integration.components.IUserPrivilegesService
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import java.net.URI
 
+@Suppress("LongParameterList")
 @Configuration
 class SubmissionWebConfig {
+
+    @Bean
+    fun collectionService(
+        collectionSqlDataService: CollectionDataService,
+        userPrivilegeService: IUserPrivilegesService,
+    ): CollectionService = CollectionService(collectionSqlDataService, userPrivilegeService)
+
+    @Bean
+    fun submitHandler(
+        submissionService: SubmissionService,
+        extSubmissionQueryService: ExtSubmissionQueryService,
+        fileSourcesService: FileSourcesService,
+        serializationService: SerializationService,
+        toSubmissionMapper: ToSubmissionMapper,
+        queryService: SubmissionMetaQueryService,
+        fileServiceFactory: FileServiceFactory,
+    ): SubmitWebHandler =
+        SubmitWebHandler(
+            submissionService,
+            extSubmissionQueryService,
+            fileSourcesService,
+            serializationService,
+            toSubmissionMapper,
+            queryService,
+            fileServiceFactory
+        )
+
+    @Bean
+    fun submissionDraftService(
+        submitWebHandler: SubmitWebHandler,
+        toSubmissionMapper: ToSubmissionMapper,
+        serializationService: SerializationService,
+        submitRequestBuilder: SubmitRequestBuilder,
+        userPrivilegesService: IUserPrivilegesService,
+        submissionQueryService: SubmissionPersistenceQueryService,
+        persistenceDraftService: SubmissionDraftPersistenceService,
+    ): SubmissionDraftService =
+        SubmissionDraftService(
+            submitWebHandler,
+            toSubmissionMapper,
+            serializationService,
+            submitRequestBuilder,
+            userPrivilegesService,
+            submissionQueryService,
+            persistenceDraftService,
+        )
 
     @Bean
     fun submissionHandler(
@@ -21,4 +84,12 @@ class SubmissionWebConfig {
     @Bean
     fun extPageMapper(properties: ApplicationProperties): ExtendedPageMapper =
         ExtendedPageMapper(URI.create(properties.instanceBaseUrl))
+
+    @Bean
+    fun submitRequestBuilder(
+        onBehalfUtils: OnBehalfUtils,
+    ): SubmitRequestBuilder = SubmitRequestBuilder(onBehalfUtils)
+
+    @Bean
+    fun onBehalfUtils(securityQueryService: ISecurityQueryService): OnBehalfUtils = OnBehalfUtils(securityQueryService)
 }
